+++ BE/src/diaries/diaries.controller.ts
@@ -17,7 +17,7 @@ import {
 } from "./diaries.dto";
 import { Diary } from "./diaries.entity";
 import { AuthGuard } from "@nestjs/passport";
-import { IdGuard } from "src/auth/auth.id-guard";
+import { IdGuard } from "src/auth/guard/auth.id-guard";
 
 @Controller("diaries")
 @UseGuards(AuthGuard())
